+++ sw.js
@@ -13,7 +13,7 @@
  */
 'use strict';
 
-var VERSION = 'v73';
+var VERSION = 'v74';
 var RUNTIME_CACHE_VERSION = 'v2';
 var SHELL_CACHE = 'stormscope-shell-' + VERSION;
 var TILE_CACHE = 'stormscope-tiles-' + RUNTIME_CACHE_VERSION;
@@ -445,7 +445,6 @@ self.addEventListener('fetch', function (event) {
   }
   if (isCameraMonolith(url)) {
     event.respondWith(staleWhileRevalidate(request, DATA_CACHE, event));
-    return;
     return;
   }